+++ oshi-core/src/main/java/module-info.java
@@ -2,7 +2,7 @@
  * JNA-based native implementation of the OSHI API for JDK 8+.
  * <p>
  * This module uses <a href="https://github.com/java-native-access/jna">JNA</a> for native access and supports all OSHI
- * platforms (Windows, macOS, Linux, Android, DragonFly BSD, FreeBSD, NetBSD, OpenBSD, Solaris, AIX).
+ * platforms.
  * <p>
  * Usage:
  *
@@ -23,7 +23,6 @@ module com.github.oshi {
     exports oshi.util.gpu;
     exports oshi.util.platform.mac;
     exports oshi.util.platform.unix.freebsd;
-    exports oshi.util.platform.unix.dragonflybsd;
     exports oshi.util.platform.unix.openbsd;
     exports oshi.util.platform.unix.solaris;
     exports oshi.util.platform.windows;