+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [#3350](https://github.com/oshi/oshi/pull/3350): Add AIX support to the FFM (`oshi-core-ffm`) implementation - [@dbwiddis](https://github.com/dbwiddis).
 * [#3360](https://github.com/oshi/oshi/pull/3360): Add NetBSD support to the FFM (`oshi-core-ffm`) implementation - [@dbwiddis](https://github.com/dbwiddis).
+* Add DragonFly BSD support to the FFM (`oshi-core-ffm`) implementation. With this, FFM and JNA implementations cover the same set of supported platforms - [@dbwiddis](https://github.com/dbwiddis).
 
 ##### Bug Fixes and Improvements